% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auctime-package.R
\docType{package}
\name{auctime-package}
\alias{auctime}
\alias{auctime-package}
\title{auctime: Calculate Incremental Area Under the Curve}
\description{
Calculate an incremental trapezoidal area under the curve for biomarkers measured at a series of time points, with the first timepoint being a baseline measurement. This is a summary measure that captures the total magnitude of a response over a specified time interval; it is an estimate of total concentration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/scrs-msu/auctime}
  \item Report bugs at \url{https://github.com/scrs-msu/auctime/issues}
}

}
\keyword{internal}
