% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAUC.R
\name{ribbonize}
\alias{ribbonize}
\title{ribbonize}
\usage{
ribbonize(.data, .x, .y, .f)
}
\arguments{
\item{.x}{x-coordinates of a function, for two functions to shade between}

\item{.y}{y-coordinates of a function, for two functions to shade between}

\item{.f}{a two-level factor, indicates whether a set of x- and y- coordinates belongs to function "a" or function "b"}
}
\value{
a tibble with 4 variables: x, x-coordinates and interpolated points for both functions; ymax, corresponding y-coordinates for the upper-bound of shading; ymin, corresponding y-coordinates for the lower-bound of shading; fill, a logical variable to indicate if an interval is shaded or not
}
\description{
calculate the ribbons required for geom_ribbon() to shade between two lines
}
\examples{
df <- tibble(
x = c(1:8, 1:8),
y = c(1, 5, 6, 4, 1, 1, 3, 2, 1, 4, 5, 4, 2, 2, 2, 2),
f = c(rep("a", 8), rep("b", 8)))

ribbons <- ribbonize(df, x, y, f)

ggplot(df) +
geom_line(aes(x, y, linetype = f)) +
geom_ribbon(data = ribbons, aes(x, ymin = ymin, ymax = ymax, fill = fill))

}
\author{
Neal Grantham, <https://www.nsgrantham.com/fill-between-two-lines-ggplot2>
}
